% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{r2_squ}
\alias{r2_squ}
\title{r2_squ}
\usage{
r2_squ(obs, res)
}
\arguments{
\item{obs}{: vector, observed values}

\item{res}{residual (observed values-predicted values)}
}
\value{
double for rsquared
}
\description{
This function is to calculate rsquared value for regression models.
}
