% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{rmse}
\alias{rmse}
\title{rmse}
\usage{
rmse(obs, pre)
}
\arguments{
\item{obs}{vector, observed values}

\item{pre}{vector, (predicted values)}
}
\value{
double for rmse
}
\description{
This function is to calculate rsquared value for regression models.
}
